messaging_app/messaging_app/requirements.txt
django
djangorestframework
django-filter
djangorestframework-simplejwt
drf-nested-routers
